# Conflicts:
#	build.gradle
#	src/main/java/net/geant/nmaas/kubernetes/KubernetesConnector.java
#	src/main/java/net/geant/nmaas/nmservice/configuration/gitlab/GitLabConfigHandler.java
#	src/main/java/net/geant/nmaas/portal/api/market/AppInstanceController.java
#	src/main/java/net/geant/nmaas/portal/api/security/CustomAccessTokenController.java
#	src/main/java/net/geant/nmaas/portal/service/impl/ApplicationBaseServiceImpl.java
#	src/main/resources/log4j2.json

+++ src/main/java/net/geant/nmaas/nmservice/configuration/GitConfigHandler.java
@@ -24,10 +24,10 @@ public interface GitConfigHandler {
      */
     AppConfigRepositoryAccessDetails configRepositoryAccessDetails(Identifier deploymentId);
 
-    void addMemberToProject(Integer gitLabProjectId, Integer gitLabUserId);
-    void addMemberToProject(Integer gitLabProjectId, String username);
+    void addMemberToProject(Long gitLabProjectId, Long gitLabUserId);
+    void addMemberToProject(Long gitLabProjectId, String username);
 
-    void removeMemberFromProject(Integer gitLabProjectId, Integer gitLabUserId);
-    void removeMemberFromProject(Integer gitLabProjectId, String username);
+    void removeMemberFromProject(Long gitLabProjectId, Long gitLabUserId);
+    void removeMemberFromProject(Long gitLabProjectId, String username);
 
 }
